Implementing an option for fast spectral resampling using numpy.interp.
The default still uses spectres for exact, flux-conserving resampling. However, if the model does not contain features narrower than the observed wavelength resolution, this may be overkill. In such cases, linear interpolation may be good enough. This is now an option for Spectrum objects using numpy.interp as a drop-in replacement for spectres. In initial tests, this is approximately 3 times faster than spectres.