Now enabling automatic creation of potential starting points by sampling from the prior, using the prior transform
This had been partially implemented and left disabled in the last commit. However, it emerged that the problem was in one of the prior transforms, rather than in the approach. It's still not a super-effective approach, because you can end up a long way from the MAP. However, with *very large* numbers of walkers, this will be an effective way to ensure the entire prior volume is sampled early on; longer chains will be required to ensure good convergence, however