Hyperion class initial revision
Starting to put the Hyperion model class into the same structure as the other model types in Ampere - broken up the __init__ call to separate functions that add a source, dust component(s), and specify the model properties. Need to allow it to deal with multiple components iteratively (in a for loop in the script that calls this model class, or internally?).