Streamlining Ensemble MCMC classes
Moving parts of ensemble MCMC around to make it easier to add new ones. So far, core functionality has been moved to parent classes. Next stage is to do the same with optimisation and post-processing, and enable the use of arviz tools for post-processing. This will make it easy to add zeus and ptemcee as samples, as well as to diversify the range of MCMC (and other) samplers on offer