Starting simulator implementation and adding ModelResults objects
This commit has a number of changes. First, a ModelResults object has been implemented.
These are intended to provide a clean way to handle a large range of types of
observations and simulations. Instead of storing results in properties of the model,
the user can pass dictionaries to a ModelResults object covering different Data types,
with @cachedproperty methods to create one type from a more complex one (e.g. a spectrum
from a data cube). How to extend these objects for new data types is still an open
question. The name ModelResults was chosen in case in future we want a results object
for inference results as well. Hopefully this will open more routes for easy caching of
models as well.

Secondly, this commit starts to implement methods for simulating results. These will
be used by simulation-based (or likelihood-free) inference packages for the most
expensive models. Eventually, amortised inference will be implemented. This results in
new inference classes in infer/sbi.py.