Lost track of changes, so there are several updates here.
Firstly, the issue with mis-reported autocorrelation times from emcee is fixed. There is still some work to do in making sure this propagates correctly to any future MCMC samplers that get added.
Secondly, performing inference with more than one approach in the same script will no longer result in duplicate logs. The logging setup first checks if the logger is already set up and if the desired handlers already exist. However, the current implementation means that you cannot change the output file for the logs between inferences. The ampere logger also no longer propagates to the root logger, which ensures that messages are not duplicated when other packages in the same script rely on the root logger (e.g. zeus).
Finally, an option is now exposed to pass naming information to the inference objects. If no name is passed, it will use the current time and the model class name to construct a name. This name is used as a prefix for the plots. In future this may be extended to the logging file.