Big change in ampere! This includes **breaking changes**!
Ampere now uses pyphot official, instead of Sundar's fork of it. This has necessitated a few significant changes.
1. Models must now return a dictionary of results, or a ModelResults object. See minimal_working_example_sbi.py for an example of how this works
2. Photometry can now accept a diverse range of inputs in different units, which will be automatically converted to the correct system internally (I hope to implement the same thing for spectra soon)
3. The minimal_working_examples have been updated to reflect these changes (examples_for_paper coming soon)!

However, these changes are not yet documented, and there seems to be an outstanding quirk with emcee that I need to work out - it prints a lot of numbers to the terminal when zeus doesn't. Still not sure what is triggering this.