Some fixes and improvements for SBI
SBI's own transforms are now disabled, this seems to prevent a number of problems with drawing samples in constrained vs unconstrained space, making the code more reliable. This was a particularly irritating crash as it would happen after the posterior was trained, hence wasting all the work that had been done.
Multi-round SNPE is now possible. This can significantly reduce the number of model evaluations required to get a given precision in the CIs, or provide a much more refined posterior with the same number of model evaluations. SBI_SNPE.optimise() now exposes an n_rounds keyword, which specifies how many rounds of inference should be performed. Each round has nsamples model evaluations