Cosmetic updates
I am making batches of cosmetic changes to keep the linters happy. This was originally going to be one giant commit of purely cosmetic changes, but in the process a few other changes have snuck in. For example:
    1. I had to run a test on a windows machine, and discovered os.path.normpath. I recommend switching all internal path logic over to use this method to ensure cross-platform compatibility
    2.  I found that minimal_working_example.py hadn't been updated to reflect the change in location of the test data
    3. Docstrings were updated in a few places

I will continue to make these sorts of cosmetic changes for the purposes of cleaner linting while making other updates.