First attempt at exposing embedding networks for sbi
Embedding networks can be useful for high-dimensional datasets with small numbers of free parameters. The idea is to train a neural network which tries to compress the data to a more compact representation (an 'embedding'; really this is just a fancy term for a subspace with reduced dimension that optimally reconstructs the data for a given set of hyperparameters, i.e. *all this is doing is dimensionality reduction*).

SBI now ships with a few default embeddings, so these are exposed, as is the means to supply your own network provided it is written with Torch.