Adding many scipy optimisers
This commit adds all of SciPy's *global* optimisers _except_ brute, because that doesn't seem like a useful solution to this problem.

Copilot did a lot of the heavy lifting in the last couple of optimisers, and I've never used them myself so I couldn't check them in as much detail as the first few.