Better automatic limits for posterior-predictive plots.
In most cases, we only care about comparing models to data in the range of the data, so the default plot range now only shows the data. This should make posterior predictive plots a bit easier to interpret without tweaking. The line for the MAP solution is now plotted in luminous magenta, so it should be easier to see, and photometry symbols are larger for the same reason.