Test splitting corner plots by models and data
Once you have many parameters, corner plots become *very* difficult to read. This commit tests and approach that splits the corner plots into two if there are more than 10 parameters. The first plot shows only the model parameters, and the second shows all the data parameters. Ideally there would be cross-plots too, but that's way too much effort for an initial test. For now it is only implemented for SBI so we can see how it does and tweak it.